64bit patches
Jaqms patches for 64bit compatibility
Fix RS97 Makefile
Add BUILDROOT makefile template

+++ sources/Application/Instruments/SampleInstrument.cpp
@@ -809,7 +809,7 @@ bool SampleInstrument::Render(int channel,fixed *buffer,int size,bool updateTick
         {
 	        if (useDirtyDownsampling_)
 	        {
-	          i1 =(short *)(((unsigned int)input)&dsMask);
+	          i1 =(short *)(((long)input)&dsMask);
 	        }
           else
 	        {
